move ray strategy to nano/deps/ray
