+++ python/nano/src/bigdl/nano/deps/ray/ray_distributed.py
@@ -30,27 +30,35 @@
 # limitations under the License.
 
 
-from typing import Callable, Dict, List, Union, Any, Optional
-
 import os
+import warnings
+from typing import Callable, Dict, List, Union, Any, Optional
 from collections import defaultdict
-from bigdl.nano.deps.ipex.ipex_api import ipex_device, ipex_optimize
-
-from bigdl.nano.pytorch.utils import TORCH_VERSION_LESS_1_10
 
 import ray
+from ray.util.ml_utils.util import find_free_port
+
 import torch
-from pytorch_lightning.plugins import DDPSpawnPlugin
-from pytorch_lightning import _logger as log, LightningModule
+from torch import Tensor
+import pytorch_lightning as pl
+from pytorch_lightning.core.datamodule import LightningDataModule
+from pytorch_lightning.trainer.states import TrainerFn
 from pytorch_lightning.utilities import rank_zero_only
-from pytorch_lightning.utilities.seed import reset_seed
-from ray.util.sgd.utils import find_free_port
-from bigdl.nano.deps.ray.ray_envbase import RayEnvironment
+from pytorch_lightning.utilities.apply_func import move_data_to_device
+from pytorch_lightning.utilities.distributed import ReduceOp
+from pytorch_lightning.strategies.launchers import _SpawnLauncher
+from pytorch_lightning.strategies import DDPSpawnStrategy
+
+from .ray_envbase import RayEnvironment
 from bigdl.nano.utils.log4Error import invalidInputError
-import warnings
+from bigdl.nano.pytorch.strategies.ipex.ipex_api import ipex_device, ipex_optimize, \
+    create_IPEXAccelerator, to_cpu
+from bigdl.nano.pytorch.utils import TORCH_VERSION_LESS_1_10
+
+_STEP_OUTPUT_TYPE = Union[torch.Tensor, Dict[str, Any]]
 
 
-@ray.remote
+@ray.remote     # type: ignore
 class RayExecutor:
     """A class to execute any arbitrary function remotely."""
 
@@ -61,60 +69,112 @@ class RayExecutor:
             os.environ[key] = value
 
     def set_env_vars(self, keys: List[str], values: List[str]):
-        """Sets multiple env vars with the provided values"""
+        """Sets multiple env vars with the provided values."""
         invalidInputError(len(keys) == len(values),
                           "keys length doesn't mathcc values length")
         for key, value in zip(keys, values):
             self.set_env_var(key, value)
 
     def get_env_vars(self, key: str):
+        """Return the specified environment variable."""
         return os.environ[key]
 
     def get_node_ip(self):
         """Returns the IP address of the node that this Ray actor is on."""
-        return ray.services.get_node_ip_address()
+        return ray.util.get_node_ip_address()
 
     def execute(self, fn: Callable, *args, **kwargs):
         """Execute the provided function and return the result."""
         return fn(*args, **kwargs)
 
 
-class RayPlugin(DDPSpawnPlugin):
-    """Pytorch Lightning plugin for DDP training on a Ray cluster.
-    This plugin is used to manage distributed training using DDP and
-    Ray for process launching. Internally, the specified number of
-    Ray actors are launched in the cluster and are registered as part of a
-    Pytorch DDP process group. The Pytorch Lightning trainer is instantiated
-    on the driver and sent to each of these training workers where training is
-    executed. The distributed training protocol is handled by Pytorch DDP.
-    Each training worker is configured to reserve ``num_cpus_per_worker``
-    CPUS and 1 GPU if ``use_gpu`` is set to ``True``.
-    If using this plugin, you should run your code like a normal Python
-    script: ``python train.py``, and only on the head node if running in a
-    distributed Ray cluster. There is no need to run this script on every
-    single node.
-    Args:
-        num_workers (int): Number of training workers to use.
-        num_cpus_per_worker (int): Number of CPUs per worker.
-        use_gpu (bool): Whether to use GPU for allocation. For GPU to be
-            used, you must also set the ``gpus`` arg in your Pytorch Lightning
-            Trainer to a value > 0.
-        init_hook (Callable): A function to run on each worker
-            upon instantiation.
-        **ddp_kwargs: Additional arguments to pass into
-            ``DistributedDataParallel`` initialization
-    Example:
-        .. code_block:: python
-            import pytorch_lightning as ptl
-            from ray_lightning import RayAccelerator
-            ptl_model = MNISTClassifier(...)
-            plugin = RayPlugin(num_workers=4, cpus_per_worker=1,
-                use_gpu=True)
-            # If using GPUs, set the ``gpus`` arg to a value > 0.
-            # The actual number of GPUs is determined by ``num_workers``.
-            trainer = pl.Trainer(..., gpus=1, plugins=[plugin])
-            trainer.fit(ptl_model)
-    """
+class _RayLauncher(_SpawnLauncher):
+
+    def __init__(self, strategy: 'RayStrategy') -> None:
+        self._strategy: RayStrategy = strategy
+
+    @property
+    def is_interactive_compatible(self) -> bool:
+        return False
+
+    def launch(self, function: Callable, *args: Any,
+               trainer: Optional["pl.Trainer"] = None, **kwargs: Any) -> Any:
+        # pytorch_lightning 1.6 uses this method to create child processes
+        invalidInputError(trainer is not None and self._strategy.cluster_environment is not None,
+                          'strategy.cluster_environment and trainer cannot be None')
+        invalidInputError(self._strategy.model is not None, "You must specify the model.")
+
+        strategy = self._strategy
+
+        # fix bug, see ddp_spawn strategy for details
+        if strategy.use_ipex and TORCH_VERSION_LESS_1_10:
+            if isinstance(args[1], LightningDataModule):
+                args[1].trainer = None
+            elif isinstance(args[3], LightningDataModule):
+                args[3].trainer = None
+            elif isinstance(args[4], LightningDataModule):
+                args[4].trainer = None
+
+        strategy._setup_env_vars()
+        strategy.global_to_local = strategy.get_local_ranks()   # type: ignore
+
+        torch_backend = os.getenv("PL_TORCH_DISTRIBUTED_BACKEND")
+        if torch_backend is None:
+            torch_backend = "nccl" if strategy.use_gpu else "gloo"
+        strategy._process_group_backend = torch_backend
+
+        futures = [
+            strategy.workers[i].execute.remote(
+                self._wrapping_function,
+                (i, trainer, function, args, kwargs)
+            )
+            for i in range(strategy.num_workers)
+        ]
+
+        results = ray.get(futures)  # type: ignore
+
+        # Get the results, checkpoint path, and model weights from worker 0.
+        results, best_path, state_dict = results[0]  # type: ignore
+        strategy._model.load_state_dict(state_dict)
+        strategy.lightning_module.trainer = trainer
+        strategy.lightning_module.trainer.checkpoint_callback.best_model_path = best_path
+
+        return results
+
+    @staticmethod
+    def _wrapping_function(args_pack: tuple) -> Any:   # type: ignore[override]
+        global_rank, trainer, function, args, kwargs = args_pack
+        strategy = trainer.strategy
+        invalidInputError(isinstance(strategy, RayStrategy), "expect ray strategy here")
+        invalidInputError(isinstance(strategy.cluster_environment, RayEnvironment),
+                          "expect ray environment here")
+
+        strategy.cluster_environment.set_global_rank(global_rank)
+        strategy.cluster_environment.set_remote_execution(True)
+
+        strategy._worker_setup(global_rank)
+        results = function(*args, **kwargs)
+
+        if trainer is not None:
+            results = strategy._launcher._collect_rank_zero_results(trainer, results)
+
+        if strategy.global_rank == 0:
+            if trainer.checkpoint_callback is not None:
+                return move_data_to_device(results, "cpu"), \
+                    trainer.checkpoint_callback.best_model_path, \
+                    move_data_to_device(strategy.lightning_module.state_dict(), "cpu")
+            else:
+                return move_data_to_device(results, "cpu"), \
+                    None, \
+                    move_data_to_device(strategy.lightning_module.state_dict(), "cpu")
+        else:
+            return None
+
+
+class RayStrategy(DDPSpawnStrategy):
+    """A DDP Strategy which uses ray as backend."""
+
+    strategy_name = "ray"
 
     def __init__(self,
                  num_workers: int = 1,
@@ -123,8 +183,8 @@ class RayPlugin(DDPSpawnPlugin):
                  use_ipex: bool = False,
                  enable_bf16: bool = False,
                  init_hook: Callable = None,
-                 **ddp_kwargs: Union[Any, Dict[str, Any]]):
-
+                 **ddp_kwargs: Any):
+        """Create a RayStrategy."""
         # Unset MKL setting as bigdl.nano would give default values when init env.
         # Running different programs may need different configurations.
         # refer to https://analytics-zoo-doc.readthedocs.io/en/master/zoo.ray.html#
@@ -134,26 +194,37 @@ class RayPlugin(DDPSpawnPlugin):
         os.environ.pop("KMP_AFFINITY", None)
         os.environ.pop("OMP_NUM_THREADS", None)
 
-        if not ray.is_initialized():
-            print(ray.init())
+        if not ray.is_initialized():    # type: ignore
+            print(ray.init())   # type: ignore
 
-        super().__init__(
-            sync_batchnorm=False,
-            parallel_devices=[],
-            cluster_environment=RayEnvironment(world_size=num_workers),
-            **ddp_kwargs)  # type: ignore
-        self.nickname = "ddp_ray"
+        if use_ipex and TORCH_VERSION_LESS_1_10 and 'accelerator' not in ddp_kwargs:
+            super().__init__(accelerator=create_IPEXAccelerator(),
+                             parallel_devices=[],
+                             cluster_environment=RayEnvironment(world_size=num_workers),
+                             **ddp_kwargs)
+        else:
+            super().__init__(parallel_devices=[],
+                             cluster_environment=RayEnvironment(world_size=num_workers),
+                             **ddp_kwargs)
+
         self.num_workers = num_workers
         self.num_cpus_per_worker = num_cpus_per_worker
         self.use_gpu = use_gpu
         self.use_ipex = use_ipex
+        self.enable_bf16 = enable_bf16
 
         invalidInputError(not self.use_gpu or not self.use_ipex,
                           "You can not specify gpu and ipex at the same time.")
 
-        self.workers: List[Any] = []
+        self.workers = self._create_worker()
         self.init_hook = init_hook
         self._local_rank = 0
+
+        if self.init_hook:
+            ray.get([w.execute.remote(self.init_hook) for w in self.workers])   # type: ignore
+
+    def _configure_launcher(self):
+        self._launcher = _RayLauncher(self)
 
     def _create_worker(self):
         """Creates Ray actor."""
@@ -163,7 +234,6 @@ class RayPlugin(DDPSpawnPlugin):
 
         workers = []
         for i in range(self.num_workers):
-
             worker = RayExecutor.options(
                 num_cpus=self.num_cpus_per_worker,
                 num_gpus=int(self.use_gpu)
@@ -176,22 +246,20 @@ class RayPlugin(DDPSpawnPlugin):
 
         return workers
 
-    def setup(self, model):
-        """Sets up PTL Trainer and creates the Ray actors."""
-        # Check that trainer attribute has been set when this method is called.
-        self._model = model
-        self.workers = self._create_worker()
-        if self.init_hook:
-            ray.get([w.execute.remote(self.init_hook) for w in self.workers])
+    def _setup_env_vars(self):
+        # Get rank 0 worker address and port for DDP connection.
+        os.environ["MASTER_ADDR"] = ray.get(
+            self.workers[0].get_node_ip.remote())
+        os.environ["MASTER_PORT"] = str(
+            ray.get(self.workers[0].execute.remote(find_free_port)))
 
-    def __getstate__(self):
-        d = self.__dict__.copy()
-        del d["workers"]
-        return d
-
-    def __setstate__(self, d):
-        d["workers"] = []
-        self.__dict__.update(d)
+        # Set environment variables for remote workers.
+        keys = [
+            "PL_GLOBAL_SEED", "PL_TORCH_DISTRIBUTED_BACKEND",
+            "MASTER_ADDR", "MASTER_PORT"
+        ]
+        values = [os.getenv(k) for k in keys]
+        ray.get([w.set_env_vars.remote(keys, values) for w in self.workers])
 
     def get_local_ranks(self):
         """Creates a mapping of global ranks to local ranks."""
@@ -206,229 +274,58 @@ class RayPlugin(DDPSpawnPlugin):
             rank_counter_dict[ip] += 1
         return global_to_local
 
-    def _setup_env_vars(self):
-        # Get rank 0 worker address and port for DDP connection.
-        os.environ["MASTER_ADDR"] = ray.get(
-            self.workers[0].get_node_ip.remote())
-        os.environ["MASTER_PORT"] = str(
-            ray.get(self.workers[0].execute.remote(find_free_port)))
-
-        # Set environment variables for remote workers.
-        keys = [
-            "PL_GLOBAL_SEED", "PL_TORCH_DISTRIBUTED_BACKEND", "MASTER_ADDR",
-            "MASTER_PORT"
-        ]
-        values = [os.getenv(k) for k in keys]
-        ray.get([w.set_env_vars.remote(keys, values) for w in self.workers])
-
-    def execution_loop(self, trainer, tune_enabled: bool = True):
-        """Main execution loop for training, testing, & prediction.
-        Sets up the torch.distributed process group for each
-        worker. Then trigger remote training/testing/eval via
-        ``train_remote`` on each worker. If using with Ray Tune, create a
-        communication queue to retrieve intermediate results, and process
-        those results. Finally retrieve the training results from the rank 0
-        worker and return."""
-
-        # Sets environment variables for all workers.
-        self._setup_env_vars()
-
-        self.global_to_local = self.get_local_ranks()
-
-        model = self._model  # type: ignore
-        model_ref = ray.put(model)
-        # Don't pickle the model when training remotely.
-
-        self._model = None  # type: ignore
-
-        futures = [
-            self.workers[i].execute.remote(self.execute_remote, model_ref, i)
-            for i in range(self.num_workers)
-        ]
-
-        not_ready = futures
-        while not_ready:
-            ready, not_ready = ray.wait(not_ready, timeout=0)
-            ray.get(ready)
-        ray.get(ready)
-
-        results = ray.get(futures)
-
-        # Get the results, checkpoint path, and model weights from worker 0.
-        results, best_path, state_dict = results[0]
-        # Set the state for PTL using the output from remote training.
-        self._results = results
-
-        self._model = model  # type: ignore
-        self._model.load_state_dict(state_dict)  # type: ignore
-        if self.lightning_module.trainer.checkpoint_callback:
-            self.lightning_module.trainer.checkpoint_callback \
-                .best_model_path = best_path
-
-        return results
-
-    def start_training(self, trainer):
-        results = self.execution_loop(trainer, tune_enabled=True)
-        # reset optimizers, since main process is never used for training and
-        # thus does not have a valid optim state.
-        trainer.optimizers = []
-        return results
-
-    def start_testing(self, trainer):
-        results = self.execution_loop(trainer, tune_enabled=False)
-        return results
-
-    def start_predicting(self, trainer):
-        results = self.execution_loop(trainer, tune_enabled=False)
-        return results
-
-    def post_dispatch(self):
-        """Shutdown the DDP process group and all the Ray actors. """
-
-        def shutdown_remote():
-            torch.distributed.destroy_process_group()
-            if torch.cuda.is_available():
-                torch.cuda.empty_cache()
-
-        ray.get([w.execute.remote(shutdown_remote) for w in self.workers])
-        for w in self.workers:
-            ray.kill(w, no_restart=True)
-            del w
-        self.workers = []
-
-    # All methods below are only executed in remote Ray workers.
-
-    def execute_remote(self,
-                       model: LightningModule,
-                       global_rank: int
-                       ):
-        """Train/test/eval function to be executed on each remote worker."""
-        invalidInputError(isinstance(self, RayPlugin), "expect ray plugin here")
-        # This method should be executed remotely in each worker.
-        self._model = model  # type: ignore
-        self.lightning_module.trainer.accelerator_connector\
-            ._training_type_plugin = self
-        self.lightning_module.trainer.accelerator.training_type_plugin = self
-
-        invalidInputError(isinstance(self.cluster_environment, RayEnvironment),
-                          "expect ray environment here")
-        if not isinstance(self.cluster_environment, RayEnvironment):
-            return
-        self.cluster_environment.set_global_rank(global_rank)
-        self.cluster_environment.set_remote_execution(True)
-
-        # Calling new_process will call
-        # transfer_distrib_spawn_state_on_fit_end.
-        # We override that method and have it just set attributes.
-        # Then we can just return those attributes here.
-        self.new_process(
-            process_idx=global_rank,
-            trainer=self.lightning_module.trainer,
-            mp_queue=None)
-        # Only need results from worker 0.
-        if self.global_rank == 0:
-            return self.results, self.best_model_path, self.model_state_dict
-        else:
-            return None
-
-    def init_ddp_connection(self,  # type: ignore
-                            global_rank: int,
-                            world_size: int,
-                            is_slurm_managing_tasks: bool = False) -> None:  # type: ignore
-        """Process group creation to be executed on each remote worker."""
-        torch_backend = os.getenv("PL_TORCH_DISTRIBUTED_BACKEND")
-        if torch_backend is None:
-            torch_backend = "nccl" if self.use_gpu else "gloo"
-
-        if not torch.distributed.is_initialized():
-            log.info(f"initializing ddp: GLOBAL_RANK: {global_rank}, MEMBER:"
-                     f" {global_rank + 1}/{world_size}")
-            torch.distributed.init_process_group(
-                backend=torch_backend,
-                rank=global_rank,
-                world_size=world_size,
-            )
-
     def set_world_ranks(self, process_idx: int = 0):
         """Set the appropriate rank attribues for the trainer."""
-        invalidInputError(
-            self.cluster_environment is not None and isinstance(self.cluster_environment,
-                                                                RayEnvironment),
-            "expect ray environment here")
-        if not (self.cluster_environment is not None and isinstance(self.cluster_environment,
-                                                                    RayEnvironment)):
-            return
-        if self.cluster_environment.is_remote():
-            self._local_rank = self.global_to_local[self.global_rank]
+        invalidInputError(self.cluster_environment is not None and isinstance(
+            self.cluster_environment, RayEnvironment), "expect ray environment here")
+        if self.cluster_environment.is_remote():    # type: ignore
+            self._local_rank = self.global_to_local[self.global_rank]   # type: ignore
             self.cluster_environment.set_global_rank(self.global_rank)
             self.cluster_environment.set_world_size(self.num_workers)
             rank_zero_only.rank = self.cluster_environment.global_rank()  # type: ignore
 
-    def new_process(self, process_idx, trainer, mp_queue):
-        self.mp_queue = mp_queue
+    def setup(self, trainer: "pl.Trainer") -> None:
+        """Setup the distributed environment of ray executor, we add ipex optimization here."""
+        self.accelerator.setup(trainer)
 
-        reset_seed()
+        trainer_fn = trainer.state.fn
+        if trainer_fn == TrainerFn.FITTING:
+            if self._layer_sync:
+                self.model = self._layer_sync.apply(self.model)
 
-        self.set_world_ranks(process_idx)
+        self.setup_precision_plugin()
 
-        # set warning rank
-        rank_zero_only.rank = self.global_rank  # type: ignore
-
-        # set up server using proc 0's ip address
-        # try to init for 20 times at max in case ports are taken
-        # where to store ip_table
-        self.init_ddp_connection(self.global_rank, self.world_size)
-
-        # TODO: we moved it to the trainer.fit after calling pre_dispatch
-        #   ... need to double check that it is the correct place
-        # self.trainer.call_setup_hook(self.model)
-
-        # on world_size=0 let everyone know training is starting
-        if self.is_global_zero and not torch.distributed.is_initialized():
-            log.info("-" * 100)
-            log.info(f"distributed_backend={self.distributed_backend}")
-            log.info(f"All DDP processes registered. Starting ddp with"
-                     "{self.world_size} processes")
-            log.info("-" * 100)
-
-        # set the ranks and devices
-        self.dist.rank = self.global_rank
-        self.dist.device = self.root_device
+        # `configure_ddp` will create a `DistributedDataParallel`, which has no
+        # `test_step` method in pytorch_lightning 1.6, which causes error when
+        # calling `trainer.test()`, so we call `configure_ddp` only when fitting
+        if trainer_fn == TrainerFn.FITTING:
+            self.configure_ddp()
+        else:
+            # `trainer.test()` won't set `model.training` to `False` automatically in pl 1.6,
+            # then the following `ipex_optimize()` call will report an error,
+            # so we need to set it to `False` manuallay
+            self.model.eval()
 
         if self.use_ipex and not TORCH_VERSION_LESS_1_10:
             dtype = torch.bfloat16 if self.enable_bf16 else None
-            num_optimizers = len(self.lightning_module.trainer.accelerator.optimizers)
+            num_optimizers = len(self.optimizers)
+
             if num_optimizers == 1:
-                optimizer = self.lightning_module.trainer.accelerator.optimizers[0]
-                ipex_optimize(self.model, optimizer=optimizer,
-                              inplace=True, dtype=dtype)
+                optimizer = self.optimizers[0]
+                ipex_optimize(self.model, optimizer=optimizer, inplace=True, dtype=dtype)
             elif num_optimizers == 0:
                 ipex_optimize(self.model, inplace=True, dtype=dtype)
             else:
                 warnings.warn(f"IPEX currently only support single optimizers, "
                               f"but got {num_optimizers}. Skip IPEX")
 
-        if self.sync_batchnorm:
-            self.model = self.configure_sync_batchnorm(self.model)
-
-        self.configure_ddp()
-
-        # Move this line here so that we can temporarily use cpu while configuring ddp
-        # and use ipex.DEVICE later on
-        # move the model to the correct device
-        #
-        # The reason for this movement is relate to unstorage tensor for ipex.
-        # So maybe another way is replacing torch.save like ipexaccelerator does.
+        # some operations in `configure_ddp` do not support XPU,
+        # which is used by ipex==1.9, so we move this line here
         self.model_to_device()
-
-        self.barrier()
-        results = trainer.run_stage()
-
-        # persist info in ddp_spawn
-        self.transfer_distrib_spawn_state_on_fit_end(results)
 
     @property
     def root_device(self):
+        """Return the root device."""
         if self.use_ipex and TORCH_VERSION_LESS_1_10:
             # Add ipex option.
             return torch.device(ipex_device())
@@ -436,30 +333,11 @@ class RayPlugin(DDPSpawnPlugin):
             return torch.device("cpu")
 
     def determine_ddp_device_ids(self):
+        """Return the index of root device."""
         # For ipex case, we also should not return any optional device id.
         if self.root_device.type == "cpu" or self.root_device.type == "xpu":
             return None
         return [self.root_device.index]
-
-    def transfer_distrib_spawn_state_on_fit_end(self, results):
-        """Sets the training output as attributes so it can be retrieved."""
-        if self.global_rank == 0:
-            # Save training results as attributes.
-            self._results = results
-
-            if self.use_ipex and TORCH_VERSION_LESS_1_10:
-                # unsupported Storage type for ipex
-                # Convert xpu tensor back to cpu
-                # refer to https://github.com/intel/intel-extension-for-pytorch/issues/158
-                self.lightning_module.to("cpu")
-
-            self.model_state_dict = self.lightning_module.state_dict()
-            best_model_path = None
-            if self.lightning_module.trainer.checkpoint_callback is not None:
-                best_model_path = \
-                    self.lightning_module.trainer.checkpoint_callback\
-                        .best_model_path
-            self.best_model_path = best_model_path
 
     @property
     def distributed_sampler_kwargs(self):
@@ -468,11 +346,62 @@ class RayPlugin(DDPSpawnPlugin):
             num_replicas=self.num_workers, rank=self.global_rank)
         return distributed_sampler_kwargs
 
-    @property
-    def require_distributed_sampler(self):
-        """This plugin requires a distributed sampler."""
-        return True
+    def reduce(self, tensor, group: Optional[Any] = None,   # type: ignore[override]
+               reduce_op: Union[ReduceOp, str] = "mean") -> Tensor:
+        """Reduces a tensor from several distributed processes to one aggregated tensor."""
+        # some operations in `super.reduce()` method do not support XPU,
+        # which will cause error when using ipex==1.9, however, these operations
+        # seems not necessary, so we just ignore these errors
+        try:
+            return super().reduce(tensor, group, reduce_op)
+        except Exception as _e:
+            return tensor
 
-    @property
-    def is_distributed(self):
-        return True
+    def training_step_end(self, output: _STEP_OUTPUT_TYPE) -> _STEP_OUTPUT_TYPE:
+        """
+        A hook to do something at the end of the train step.
+
+        For ipex xpu tensor do not support `tensor.storage()` right now,
+        which is a required operation by pytorch_lightning,
+        so just move output to cpu to store it, and move it back when doing backward.
+        """
+        if self.use_ipex and TORCH_VERSION_LESS_1_10:
+            output = to_cpu(output)
+
+        return super().training_step_end(output)
+
+    def test_step_end(self, output: Optional[_STEP_OUTPUT_TYPE]) -> \
+            Optional[_STEP_OUTPUT_TYPE]:
+        """
+        A hook to do something at the end of the test step.
+
+        :param output: the output of the test step
+        """
+        if self.use_ipex and TORCH_VERSION_LESS_1_10:
+            output = to_cpu(output)
+
+        return super().test_step_end(output)
+
+    def validation_step_end(self, output: Optional[_STEP_OUTPUT_TYPE]) -> \
+            Optional[_STEP_OUTPUT_TYPE]:
+        """
+        A hook to do something at the end of the validation step.
+
+        :param output: the output of the validation step
+        """
+        if self.use_ipex and TORCH_VERSION_LESS_1_10:
+            output = to_cpu(output)
+
+        return super().validation_step_end(output)
+
+    def backward(self,  # type: ignore
+                 closure_loss: torch.Tensor,
+                 *args,
+                 **kwargs) -> torch.Tensor:
+        """Moving back loss to xpu device."""
+        closure_loss = closure_loss.to(self.root_device)
+        return super().backward(
+            closure_loss,
+            *args,
+            **kwargs,
+        )
