+++ python/nano/src/bigdl/nano/pytorch/strategies/__init__.py
@@ -17,4 +17,4 @@
 from .ipex.ipex_api import create_IPEXStrategy
 from .ddp_spawn import DDPSpawnStrategy
 from .ddp_subprocess import DDPSubprocessStrategy
-from .ray.ray_api import create_RayStrategy
+from bigdl.nano.deps.ray.ray_api import create_RayStrategy
