--- python/nano/src/bigdl/nano/pytorch/strategies/ray/ray_api.py
@@ -1,20 +1,0 @@
-#
-# Copyright 2016 The BigDL Authors.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-
-def create_RayStrategy(*args, **kwargs):
-    """Create ray strategy."""
-    from .ray_distributed import RayStrategy
-    return RayStrategy(*args, **kwargs)
