+++ python/nano/test/mypy.ini
@@ -8,13 +8,13 @@ strict_optional = False
 [mypy-torchvision.*]
 ignore_missing_imports = True
 
-# disable the type check of ipex and ray deps for pl 1.4,
+# disable the type check of ipex for pl 1.4,
 # which cannot pass the type check when using pl 1.6
 [mypy-bigdl.nano.deps.ipex.*]
 ignore_errors = True
 
 [mypy-bigdl.nano.deps.ray.*]
-ignore_errors = True
+strict_optional = False
 
 [mypy-bigdl.nano.pytorch.runtime_binding.*]
 ignore_errors = True
